Add files via upload
Added the "required: True" attribute to the MDTextfield widget, this ensures that the on_release method will not process with no input, alleviating the need for an "if MDTextfield.text == "" :" check. Also, played around with the UI a bit, all in all I am happy with the app at this point.